refactor: extract bundle bake result
