+++ bpy_speckle/converter/from_bundle/_baking/__init__.py
@@ -1,0 +1,1 @@
+"""Private Blender construction internals for bundle receive."""
